refactor: Immutableサービスをリネームし全関連参照を更新
- ImmutableChainService.ts → ChainService.ts
- ImmutableEliminationService.ts → EliminationService.ts
- ImmutableGravityService.ts → GravityService.ts
- テストファイルと依存レポートをリネームに追随して修正
