+++ app/src/domain/services/EliminationService.ts
