+++ app/src/domain/services/GravityService.ts
